use strict tslint for uni-app
