+++ types/uni-app/tslint.json
@@ -1,8 +1,3 @@
 {
-    "extends": "dtslint/dt.json",
-    "rules": {
-        "max-line-length": false,
-        "no-trailing-whitespace": false,
-        "no-unnecessary-generics": false
-    }
+    "extends": "dtslint/dt.json"
 }
